beacon-planner-test/requirements.txt
allure-pytest==2.8.24
allure-python-commons==2.8.24
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
enum34==1.1.10
idna==2.10
iniconfig==1.1.1
jsonpath==0.82
lxml==4.6.2
namedlist==1.8
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
pytest-order==0.9.2
pytest-pycharm==0.7.0
requests==2.25.1
scope==0.1.1
toml==0.10.2
urllib3==1.26.2
pandas
selenium
webdriver_manager
mailosaur